APPS-1078 add Jenkinsfile, fix warnings
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@69170 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/kernel/algos/AlgoNumeratorDenominator.java
@@ -107,7 +107,7 @@ public class AlgoNumeratorDenominator extends AlgoElement {
 				long num = (long) top.evaluateDouble();
 				long den = (long) bottom.evaluateDouble();
 				long gcd = Kernel.gcd(num, den);
-				double val = (type == Commands.Numerator) ? num / gcd : den / gcd;
+				long val = (type == Commands.Numerator) ? num / gcd : den / gcd;
 
 				g.setValue(val);
 				return;
